+++ src/main/java/org/sonatype/nexus/ci/iq/IqPolicyEvaluatorWorkflowStep.groovy
@@ -116,7 +116,7 @@ class IqPolicyEvaluatorWorkflowStep
     @Override
     FormValidation doVerifyCredentials(@QueryParameter @Nullable String jobCredentialsId, @AncestorInPath Job job)
     {
-      FormUtil.validateJobCredentials(jobCredentialsId, job)
+      IqUtil.verifyJobCredentials(jobCredentialsId, job)
     }
   }
 }
